Fixed bug in the passing of redshifts to the luminosity function. Issue with float/double

+++ src/py21cmmc/_21cmfast/wrapper.py
@@ -722,7 +722,7 @@ def compute_luminosity_function(*, redshifts, user_params=None, cosmo_params=Non
     astro_params = AstroParams(astro_params)
     flag_options = FlagOptions(flag_options)
 
-    redshifts = np.array(redshifts)
+    redshifts = np.array(redshifts, dtype='float32')
 
     lfunc = np.zeros(len(redshifts) * nbins)
     Muvfunc = np.zeros(len(redshifts) * nbins)
